Speech to text conversion followed by sentiment analysis
Here,the speech was first converted into text which was further used for purpose like sentiment analysis.Sentiment analysis is the process of identifying and categorizing opinions using NLP to determine whether the user’s/speaker’s/writer’s attitude towards a particular topic/ product is positive, negative or neutral. It was done easily with TextBlob library in Python.